requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# RSS Feed parsing
feedparser>=6.0.10
pyyaml>=6.0.1

# Reddit API
praw>=7.7.1

# HTML parsing (si necesitas scraping web)
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Data processing
python-dateutil>=2.8.2

# OpenAI para generar tweets con IA
openai>=1.12.0

# Optional: Para análisis de sentimientos
# textblob>=0.17.1

# Machine Learning (optional - para modelo de selección avanzado)
numpy>=1.24.0
scikit-learn>=1.3.0

# Development tools
pytest>=7.4.3
black>=23.12.0
flake8>=6.1.0

